Changed the Instrumented to long.

+++ EDA_Project/src/pt/europeia/eda/secondDeliver/InstrumentedInsertion.java
@@ -2,10 +2,10 @@ package pt.europeia.eda.secondDeliver;
 
 public class InstrumentedInsertion {
 
-	private static int numberOfComparisons = 0;
-	private static int numberOfArrayReads = 0;
-	private static int numberOfArrayWrites = 0;
-	private static int numberOfSwaps = 0;
+	private static long numberOfComparisons = 0;
+	private static long numberOfArrayReads = 0;
+	private static long numberOfArrayWrites = 0;
+	private static long numberOfSwaps = 0;
 
 	private InstrumentedInsertion() {
 		throw new RuntimeException("Attempt to instantiate package-class");
@@ -41,23 +41,23 @@ public class InstrumentedInsertion {
 		return true;
 	}
 
-	public static int getNumberOfComparisons() {
+	public static long getNumberOfComparisons() {
 		return numberOfComparisons;
 	}
 
-	public static int getNumberOfArrayReads() {
+	public static long getNumberOfArrayReads() {
 		return numberOfArrayReads;
 	}
 
-	public static int getNumberOfArrayWrites() {
+	public static long getNumberOfArrayWrites() {
 		return numberOfArrayWrites;
 	}
 
-	public static int getNumberOfArrayAccesses() {
+	public static long getNumberOfArrayAccesses() {
 		return numberOfArrayReads + numberOfArrayWrites;
 	}
 	
-	public static int getNumberOfSwaps() {
+	public static long getNumberOfSwaps() {
 		return numberOfSwaps;
 	}
 }
